backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
websockets==12.0
opencv-python==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
